rpi/requirements.txt
pygame
pyserial
AWSIoTPythonSDK
